backend/requirements.txt
Django
djangorestframework
django-cors-headers
drf-spectacular
djangorestframework-simplejwt
psycopg2-binary
weasyprint
qrcode
pillow
whitenoise
gunicorn
uvicorn
